Make the audit log and request page show on the same dashboard page with tabs to switch between them
